etta

+++ src/main/java/com/saasdemo/backend/dto/SubscriptionDTO.java
@@ -1,11 +1,9 @@
 package com.saasdemo.backend.dto;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class SubscriptionDTO {
-   BigDecimal amount;
+   double amount;
     
 }
